move page-header component to angular folder, scss to angular.scss
