+++ public/sass/_grafana.scss
@@ -43,7 +43,6 @@
 @import 'components/row';
 @import 'components/json_explorer';
 @import 'components/dashboard_grid';
-@import 'components/page_header';
 @import 'components/add_data_source';
 @import 'components/panel_header';
