--- public/sass/components/_page_header.scss
@@ -1,105 +1,0 @@
-.page-header {
-  margin-top: $space-md;
-
-  .btn {
-    float: right;
-    margin-left: $space-md;
-
-    // better align icons
-    .fa {
-      position: relative;
-      top: 1px;
-    }
-  }
-
-  @include media-breakpoint-up(md) {
-    margin-top: 0;
-  }
-}
-
-.page-header__inner {
-  flex-grow: 1;
-  display: flex;
-  padding: $space-lg 0;
-}
-
-.page-header__title {
-  font-size: $font-size-h2;
-  margin-bottom: 1px;
-  padding-top: $spacer;
-}
-
-.page-header__img {
-  position: relative;
-  top: 10px;
-  height: 50px;
-}
-
-.page-header__icon {
-  font-size: 50px;
-  width: 50px;
-  height: 50px;
-  position: relative;
-  color: $text-color-weak;
-
-  &.fa {
-    top: 10px;
-  }
-
-  &.gicon {
-    top: 10px;
-  }
-
-  &.icon-gf {
-    top: 3px;
-  }
-}
-
-.page-header__logo {
-  margin: -1px $spacer -1px 0;
-  color: $text-color-weak;
-}
-
-.page-header__sub-title {
-  color: $text-muted;
-}
-
-.page-header-stamps-type {
-  color: $link-color-disabled;
-  text-transform: uppercase;
-}
-
-.page-header__select-nav {
-  margin-bottom: 10px;
-  max-width: 100%;
-
-  @include media-breakpoint-up(lg) {
-    display: none;
-  }
-}
-
-.page-header__link {
-  color: $text-color;
-  text-decoration: underline;
-}
-
-.page-header__tabs {
-  display: none;
-  @include media-breakpoint-up(lg) {
-    display: block;
-  }
-}
-
-.page-breadcrumbs {
-  display: flex;
-  padding: 10px 0;
-  line-height: 0.5;
-}
-
-.breadcrumb {
-  display: inline-block;
-  box-shadow: 0 0 15px 1px rgba(0, 0, 0, 0.35);
-  overflow: hidden;
-  border-radius: 5px;
-  counter-reset: flag;
-}
